Fix APDU get_data implementation.
Comm::get_data previously was reading from the same locations as p1 and
p2; this change brings get_data in line with the expected format with
the fifth byte or fifth, sixth, and seventh bytes of the APDU as length.
It also parses the ISO7816 format or 16-bit lengths, wherein byte five
is 0 and the sixth and seventh bytes make up the 16-bit length.