Allow deriving the instruction from all 4 metadata fields
This makes it easier to not forget to check those fields when using
`next_event`. Accidental ignoring of these fields has been flagged by
auditors.

I was going to pass the metadata by reference, but the borrow-checker
didn't like `T: for<'a> TryFrom<&'a ApduMeta>` even with a `+ 'static`;
plus, 4 bytes and a 32-bit pointer are the same size. I would hope all
this would get inlined so it wouldn't matter anyways.